rename and export component
