+++ src/components/site/BrandText.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 
 const LogoType = props => (
-  <div className="dg_logotype">
+  <div className="dg_brand-text">
     <span>{props.header}</span>
     <span>{props.description}</span>
   </div>
