+++ src/dotgov.scss
@@ -42,7 +42,7 @@
 @import "./styles/partials/components/button-dark";
 @import "./styles/partials/components/table";
 @import "./styles/partials/components/site-header";
-@import "./styles/partials/components/logo-type";
+@import "./styles/partials/components/brand-text";
 @import "./styles/partials/components/site-status";
 @import "./styles/partials/components/emphasized-image";
 @import "./styles/partials/components/search";
