+++ src/index.js
@@ -14,6 +14,7 @@ export { default as RadioButton } from "./components/basic/form/RadioButton";
 export { default as Select } from "./components/basic/form/Select";
 export { default as TextArea } from "./components/basic/form/TextArea";
 export { default as TextInput } from "./components/basic/form/TextInput";
+export { default as Fieldset } from "./components/basic/form/FieldSet";
 
 /** Lists */
 export { default as List } from "./components/basic/lists/List";
@@ -28,6 +29,7 @@ export {
 /** Site */
 export { default as EmphasizedImage } from "./components/site/EmphasizedImage";
 export { default as SiteHeader } from "./components/site/SiteHeader";
+export { default as BrandText } from "./components/site/BrandText";
 
 /** Other */
 export { default as Button } from "./components/basic/Button";
