+++ src/styles/partials/components/_brand-text.scss
@@ -1,4 +1,4 @@
-.dg_logotype {
+.dg_brand-text {
   border-right: $border-width-small solid $gray-light;
   padding: 0 $padding-large 0 $padding-largest;
   text-decoration: none;
